backend/requirements.txt
# Existing dependencies
flask==3.0.0
flask-cors==4.0.0
pygame==2.5.2
schedule==1.2.0
mutagen==1.47.0

# New dependencies for enhanced functionality
flask-socketio==5.3.6
python-socketio==5.10.0
eventlet==0.33.3
werkzeug==3.0.1
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
smtplib-ssl==1.0.0
email-validator==2.1.0
psutil==5.9.6
watchdog==3.0.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-asyncio==0.21.1